apply_source_init_params apply only once
