+++ tests/sources/test_base.py
@@ -9,8 +9,7 @@ import dature
 from dature import JsonSource, Source, load
 from dature.errors import EnvVarExpandError
 from dature.field_path import F
-from dature.loading.merge_config import SourceParams
-from dature.loading.source_loading import apply_source_init_params
+from dature.loading.merge_config import SourceParams, apply_source_init_params
 from dature.sources.base import FileFieldMixin
 from dature.sources.retort import string_value_loaders, transform_to_dataclass
 from dature.sources.yaml_ import Yaml12Source
